see Mantis #302 and 303
meta and metadb from herbarinput: not only the content of selected columns is copied but the complete tables including their structure are copied. So any additionally needed column can be inserted into the structure of herbarinput.meta and herbarinput.metadb.

all tables tbl_prj_gbif_pilot_% contain an additional column "thumb_url" with a link to the thumbnail of the image.